Modified the setControlValue(Object) method so that if the first element of the String array given to the method is 'null' the method does not look for the currently selected radio button but instead sets the values of the currently displayed textfields.
Also, on June 6, 2005 changes were made that were not logged.  Mainly,
the FilteredTextField inner class was made that is a JTextField with a
StringFilter on it.  The textfields used on this ViewControl were
changed to FilteredTextFields.  Furthermore, methods were added to
add and remove filters to any textfield on the ViewControl.