Added javadocs and rearranged the code to make it easier to read.
Modified the default values specified in this class so that this class,
the CompositeContourControl class, and the ContourJPanel class all
have their defaults linked.

Added controls to allow the user to enable/disable labels and label
formatting.  Added controls to allow the user to enable/disable
certain lines with their styles.